+++ src/dataops/views.py
@@ -41,7 +41,7 @@ def row_update(request):
         return redirect('workflow:index')
 
     # If the workflow has no data, something went wrong, go back to the
-    # rowfilter page
+    # dataops home
     if workflow.nrows == 0:
         return redirect('dataops:list')
 
@@ -103,7 +103,7 @@ def row_update(request):
 
     # If there is no unique key, something went wrong.
     if not unique_field:
-        raise Exception('Unique key not found when updating row')
+        raise Exception('Key value not found when updating row')
 
     pandas_db.update_row(workflow.id,
                          set_fields,