+++ src/workflow/static/workflow/js/workflow.js
@@ -63,5 +63,9 @@ $(function () {
   // Delete column
   $("#column-table").on("click", ".js-column-delete", loadForm)
   $("#modal-item").on("submit", ".js-column-delete-form", saveForm)
+
+  // Delete column
+  $("#column-table").on("click", ".js-column-clone", loadForm)
+  $("#modal-item").on("submit", ".js-column-clone-form", saveForm)
 });